requirements.txt
python-dotenv
openai
pandas
langchain
fastapi
langchain-openai
dataclasses
openpyxl
scikit-learn
black